Safari-611.1.10/LayoutTests/fast/dom/insertAdjacentHTML-DocumentFragment-parent-expected.txt

Tests that insertAdjacentHTML() works if the element's parent is a DocumentFragment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.insertAdjacentHTML('beforebegin', '<a>test</a>') did not throw exception.
PASS fragment.firstChild.tagName is "A"
PASS fragment.firstChild.nextSibling.tagName is "DIV"
PASS successfullyParsed is true

TEST COMPLETE
